refactor: 좌석 엔티티  연관관계 변경
