+++ domain/src/main/java/com/example/entity/Screening.java
@@ -7,8 +7,6 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
@@ -31,17 +29,4 @@ public class Screening extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "theater_id")
     private Theater theater;
-
-    @OneToMany(mappedBy = "screening")
-    private List<Seat> seats = new ArrayList<>();
-
-    public void initializeSeats() {
-        String[] rows = {"A", "B", "C", "D", "E"};
-        for (String row : rows) {
-            for (int i = 1; i <= 5; i++) {
-                String seatNumber = row + i;
-                this.seats.add(new Seat(seatNumber, this));
-            }
-        }
-    }
 }
