+++ domain/src/main/java/com/example/entity/Theater.java
@@ -5,6 +5,9 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -16,7 +19,20 @@ public class Theater extends BaseEntity {
     private Long id;
     private String name;
 
+    @OneToMany(mappedBy = "theater")
+    private List<Seat> seats = new ArrayList<>();
+
     public Theater(String name) {
         this.name = name;
+    }
+
+    public void initializeSeats() {
+        String[] rows = {"A", "B", "C", "D", "E"};
+        for (String row : rows) {
+            for (int i = 1; i <= 5; i++) {
+                String seatNumber = row + i;
+                this.seats.add(new Seat(seatNumber, this));
+            }
+        }
     }
 }
